Isolate effects tool registration from the server facade
The MCP server facade was still carrying visual effects, layout, mograph, and transition registrations directly, keeping server.py large after the previous extraction waves. This moves that bounded surface into server_tools_effects while preserving re-exported names from server.py so existing imports and tool registration behavior remain stable.

Constraint: Public MCP tool names and signatures must remain unchanged.

Rejected: Merge these handlers into advanced tools | effects and transitions are a distinct user-facing surface and would make that module too broad.

Confidence: high

Scope-risk: narrow

Directive: Keep server.py as a registration facade; add new tool families to focused server_tools_* modules instead of growing the facade again.

Tested: ruff check mcp_video/server.py mcp_video/server_tools_effects.py tests/test_server.py tests/test_public_surface.py

Tested: python3 -c 'import mcp_video; import mcp_video.server'

Tested: python3 -m pytest tests/test_server.py tests/test_public_surface.py -q --tb=short

Tested: MCP registration smoke listed 83 tools and all moved effects/transition tools

Tested: python3 -m pytest tests/test_effects_engine.py tests/test_transitions.py -q --tb=short

Not-tested: Full python3 -m pytest tests/ -x -q --tb=short stops at missing local real-media fixture out/McpVideoExplainer-FINAL.mp4 after 538 passed and 9 skipped
